+++ src/main/java/rabbit/umc/com/demo/converter/ArticleConverter.java
@@ -13,8 +13,8 @@ import rabbit.umc.com.demo.community.dto.ArticleListRes.ArticleDto;
 import rabbit.umc.com.demo.community.dto.ArticleRes;
 import rabbit.umc.com.demo.community.dto.ArticleRes.ArticleImageDto;
 import rabbit.umc.com.demo.community.dto.ArticleRes.CommentDto;
-import rabbit.umc.com.demo.community.dto.CommunityHomeRes.PopularArticleDto;
-import rabbit.umc.com.demo.community.dto.CommunityHomeResV2.PopularArticleDtoV2;
+import rabbit.umc.com.demo.home.dto.CommunityHomeRes.PopularArticleDto;
+import rabbit.umc.com.demo.home.dto.CommunityHomeResV2.PopularArticleDtoV2;
 import rabbit.umc.com.demo.community.dto.GetPopularArticleRes;
 import rabbit.umc.com.demo.community.dto.PostArticleReq;
 import rabbit.umc.com.demo.mainmission.domain.MainMission;
