+++ src/main/java/rabbit/umc/com/demo/converter/MainMissionConverter.java
@@ -8,8 +8,8 @@ import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import rabbit.umc.com.demo.base.Status;
 import rabbit.umc.com.demo.community.domain.Category;
-import rabbit.umc.com.demo.community.dto.CommunityHomeRes.MainMissionDto;
-import rabbit.umc.com.demo.community.dto.CommunityHomeResV2.MainMissionDtoV2;
+import rabbit.umc.com.demo.home.dto.CommunityHomeRes.MainMissionDto;
+import rabbit.umc.com.demo.home.dto.CommunityHomeResV2.MainMissionDtoV2;
 import rabbit.umc.com.demo.mainmission.domain.MainMission;
 import rabbit.umc.com.demo.mainmission.domain.mapping.LikeMissionProof;
 import rabbit.umc.com.demo.mainmission.domain.mapping.MainMissionProof;
