+++ src/main/java/rabbit/umc/com/demo/home/converter/HomeConverter.java
@@ -4,12 +4,12 @@ import java.util.List;
 import java.util.stream.Collectors;
 import rabbit.umc.com.demo.community.domain.Article;
 import rabbit.umc.com.demo.community.domain.Category;
-import rabbit.umc.com.demo.community.dto.CommunityHomeRes;
-import rabbit.umc.com.demo.community.dto.CommunityHomeRes.MainMissionDto;
-import rabbit.umc.com.demo.community.dto.CommunityHomeRes.PopularArticleDto;
-import rabbit.umc.com.demo.community.dto.CommunityHomeResV2;
-import rabbit.umc.com.demo.community.dto.CommunityHomeResV2.MainMissionDtoV2;
-import rabbit.umc.com.demo.community.dto.CommunityHomeResV2.PopularArticleDtoV2;
+import rabbit.umc.com.demo.home.dto.CommunityHomeRes;
+import rabbit.umc.com.demo.home.dto.CommunityHomeRes.MainMissionDto;
+import rabbit.umc.com.demo.home.dto.CommunityHomeRes.PopularArticleDto;
+import rabbit.umc.com.demo.home.dto.CommunityHomeResV2;
+import rabbit.umc.com.demo.home.dto.CommunityHomeResV2.MainMissionDtoV2;
+import rabbit.umc.com.demo.home.dto.CommunityHomeResV2.PopularArticleDtoV2;
 import rabbit.umc.com.demo.converter.ArticleConverter;
 import rabbit.umc.com.demo.converter.MainMissionConverter;
 import rabbit.umc.com.demo.mainmission.domain.MainMission;
