+++ src/main/java/rabbit/umc/com/demo/home/dto/CommunityHomeRes.java
@@ -1,11 +1,9 @@
-package rabbit.umc.com.demo.community.dto;
+package rabbit.umc.com.demo.home.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
-import lombok.Data;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 import java.util.List;
 
 @Getter
