+++ src/main/java/rabbit/umc/com/demo/home/facade/HomeFacade.java
@@ -9,8 +9,8 @@ import rabbit.umc.com.demo.community.article.service.ArticleService;
 import rabbit.umc.com.demo.community.category.CategoryService;
 import rabbit.umc.com.demo.community.domain.Article;
 import rabbit.umc.com.demo.community.domain.Category;
-import rabbit.umc.com.demo.community.dto.CommunityHomeRes;
-import rabbit.umc.com.demo.community.dto.CommunityHomeResV2;
+import rabbit.umc.com.demo.home.dto.CommunityHomeRes;
+import rabbit.umc.com.demo.home.dto.CommunityHomeResV2;
 import rabbit.umc.com.demo.home.converter.HomeConverter;
 import rabbit.umc.com.demo.mainmission.domain.MainMission;
 import rabbit.umc.com.demo.mainmission.service.MainMissionService;
